Add explicit step to start httpbin
