+++ tests/integration/features/istio-jwt/Istio_jwt_two_strategies.feature
@@ -1,7 +1,8 @@
 Feature: Exposing different endpoints with different handlers
 
   Scenario: Exposing httpbin endpoints with JWT and OAuth2
-    Given OAuth2: There is an endpoint secured with JWT on path "/ip" requiring scopes '["read", "write"]'
+    Given OAuth2: There is a httpbin service
+    And OAuth2: There is an endpoint secured with JWT on path "/ip" requiring scopes '["read", "write"]'
     And OAuth2: There is an endpoint secured with OAuth2 on path "/get" requiring scopes '["read", "write"]'
     When OAuth2: The APIRule is applied
     Then OAuth2: Calling the "/ip" endpoint without a token should result in status between 400 and 403
