+++ tests/integration/features/istio-jwt/istio_jwt.feature
@@ -1,13 +1,15 @@
 Feature: Exposing endpoints with Istio JWT authorization strategy
 
   Scenario: Calling a httpbin endpoint secured
-    Given Common: There is an endpoint secured with JWT on path "/ip"
+    Given Common: There is a httpbin service
+    When Common: Common: The APIRule with an endpoint secured with JWT on path /ip is applied
     Then Common: Calling the "/ip" endpoint without a token should result in status between 400 and 403
     And Common: Calling the "/ip" endpoint with an invalid token should result in status between 400 and 403
     And Common: Calling the "/ip" endpoint with a valid "JWT" token should result in status between 200 and 299
 
   Scenario: Calling httpbin that has an endpoint secured by JWT and unrestricted endpoints
-    Given JwtAndUnrestricted: There is an endpoint secured with JWT on path "/ip"
+    Given JwtAndUnrestricted: There is a httpbin service
+    And JwtAndUnrestricted: There is an endpoint secured with JWT on path "/ip"
     And JwtAndUnrestricted: There is an endpoint with handler "allow" on path "/headers"
     And JwtAndUnrestricted: There is an endpoint with handler "noop" on path "/json"
     When JwtAndUnrestricted: The APIRule is applied
@@ -16,7 +18,8 @@ Feature: Exposing endpoints with Istio JWT authorization strategy
     And JwtAndUnrestricted: Calling the "/json" endpoint without token should result in status between 200 and 299
 
   Scenario: Calling a httpbin endpoint secured with JWT that requires scopes claims
-    Given Scopes: There is an endpoint secured with JWT on path "/ip" requiring scopes '["read", "write"]'
+    Given Scopes: There is a httpbin service
+    And Scopes: There is an endpoint secured with JWT on path "/ip" requiring scopes '["read", "write"]'
     And Scopes: There is an endpoint secured with JWT on path "/get" requiring scopes '["test", "write"]'
     And Scopes: There is an endpoint secured with JWT on path "/headers" requiring scopes '["read"]'
     When Scopes: The APIRule is applied
@@ -25,7 +28,8 @@ Feature: Exposing endpoints with Istio JWT authorization strategy
     And Scopes: Calling the "/headers" endpoint with a valid "JWT" token with scope claims "read" and "write" should result in status between 200 and 299
 
   Scenario: Calling a httpbin endpoint secured with JWT that requires aud claim
-    Given Audiences: There is an endpoint secured with JWT on path "/get" requiring audiences '["https://example.com"]'
+    Given Audiences: There is a httpbin service
+    And Audiences: There is an endpoint secured with JWT on path "/get" requiring audiences '["https://example.com"]'
     And Audiences: There is an endpoint secured with JWT on path "/ip" requiring audiences '["https://example.com", "https://example.com/user"]'
     And Audiences: There is an endpoint secured with JWT on path "/headers" requiring audiences '["https://example.com", "https://example.com/admin"]'
     When Audiences: The APIRule is applied
@@ -35,14 +39,16 @@ Feature: Exposing endpoints with Istio JWT authorization strategy
 
 
   Scenario: Endpoints secured by JWT should fallback to service defined on root level when there is no service defined on rule level
-    Given ServiceFallback: There is an endpoint secured with JWT on path "/headers" with service definition
+    Given ServiceFallback: There is a httpbin service
+    And ServiceFallback: There is an endpoint secured with JWT on path "/headers" with service definition
     And ServiceFallback: There is an endpoint secured with JWT on path "/ip"
     When ServiceFallback: The APIRule with service on root level is applied
     Then ServiceFallback: Calling the "/headers" endpoint with a valid "JWT" token should result in status between 200 and 299
     And ServiceFallback: Calling the "/ip" endpoint with a valid "JWT" token should result in status between 200 and 299
 
   Scenario: Calling a httpbin endpoint secured with JWT in two namespaces
-    Given TwoNamespaces: There are two namespaces with workload
+    Given TwoNamespaces: There is a httpbin service
+    And TwoNamespaces: There are two namespaces with workload
     And TwoNamespaces: There is an endpoint secured with JWT on path "/get" in APIRule Namespace
     And TwoNamespaces: There is an endpoint secured with JWT on path "/hello" in different namespace
     When TwoNamespaces: The APIRule is applied
@@ -52,7 +58,8 @@ Feature: Exposing endpoints with Istio JWT authorization strategy
     And TwoNamespaces: Calling the "/hello" endpoint without token should result in status between 400 and 403
 
   Scenario: Exposing different services with same methods
-   Given DiffSvcSameMethods: There is a workload and service for httpbin and helloworld
+   Given DiffSvcSameMethods: There is a httpbin service
+    And DiffSvcSameMethods: There is a workload and service for httpbin and helloworld
    And DiffSvcSameMethods: There is an endpoint secured with JWT on path "/headers" for httpbin service with methods '["GET", "POST"]'
    And DiffSvcSameMethods: There is an endpoint secured with JWT on path "/hello" for helloworld service with methods '["GET", "POST"]'
    When DiffSvcSameMethods: The APIRule is applied
@@ -60,11 +67,13 @@ Feature: Exposing endpoints with Istio JWT authorization strategy
    And DiffSvcSameMethods: Calling the "/hello" endpoint with a valid "JWT" token should result in status between 200 and 299
 
   Scenario: Exposing a JWT secured endpoint with unavailable issuer and jwks URL
+    Given JwtIssuerUnavailable: There is a httpbin service
     Given JwtIssuerUnavailable: There is an endpoint secured with JWT on path "/ip" with invalid issuer and jwks
     When JwtIssuerUnavailable: The APIRule is applied
     And JwtIssuerUnavailable: Calling the "/ip" endpoint with a valid "JWT" token should result in body containing "Jwt issuer is not configured"
 
   Scenario: Exposing a JWT secured endpoint where issuer URL doesn't belong to jwks URL
-    Given JwtIssuerJwksNotMatch: There is an endpoint secured with JWT on path "/ip" with invalid issuer and jwks
+    Given JwtIssuerJwksNotMatch: There is a httpbin service
+    And JwtIssuerJwksNotMatch: There is an endpoint secured with JWT on path "/ip" with invalid issuer and jwks
     When JwtIssuerJwksNotMatch: The APIRule is applied
     And JwtIssuerJwksNotMatch: Calling the "/ip" endpoint with a valid "JWT" token should result in body containing "Jwks doesn't have key to match kid or alg from Jwt"
