+++ tests/integration/scenario_audience_test.go
@@ -15,6 +15,7 @@ func initAudience(ctx *godog.ScenarioContext) {
 
 	scenario := istioJwtManifestScenario{s}
 
+	ctx.Step(`Audiences: There is a httpbin service$`, scenario.thereIsAHttpbinService)
 	ctx.Step(`Audiences: There is an endpoint secured with JWT on path "([^"]*)" requiring audiences '(\[.*\])'$`, scenario.thereIsAnEndpointWithAudiences)
 	ctx.Step(`Audiences: The APIRule is applied$`, scenario.theAPIRuleIsApplied)
 	ctx.Step(`Audiences: Calling the "([^"]*)" endpoint with a valid "([^"]*)" token with audiences "([^"]*)" and "([^"]*)" should result in status between (\d+) and (\d+)`, scenario.callingTheEndpointWithAValidToken)
