+++ tests/integration/scenario_common_jwt_test.go
@@ -14,7 +14,8 @@ func initCommon(ctx *godog.ScenarioContext) {
 
 	scenario := istioJwtManifestScenario{s}
 
-	ctx.Step(`Common: There is an endpoint secured with JWT on path "([^"]*)"$`, scenario.theAPIRuleIsApplied)
+	ctx.Step(`Common: There is a httpbin service$`, scenario.thereIsAHttpbinService)
+	ctx.Step(`Common: The APIRule with an endpoint secured with JWT on path /ip is applied$`, scenario.theAPIRuleIsApplied)
 	ctx.Step(`Common: Calling the "([^"]*)" endpoint without a token should result in status between (\d+) and (\d+)$`, scenario.callingTheEndpointWithoutTokenShouldResultInStatusBetween)
 	ctx.Step(`Common: Calling the "([^"]*)" endpoint with an invalid token should result in status between (\d+) and (\d+)$`, scenario.callingTheEndpointWithInvalidTokenShouldResultInStatusBetween)
 	ctx.Step(`Common: Calling the "([^"]*)" endpoint with a valid "([^"]*)" token should result in status between (\d+) and (\d+)$`, scenario.callingTheEndpointWithValidTokenShouldResultInStatusBetween)
