+++ tests/integration/scenario_jwt_and_unrestricted_test.go
@@ -14,6 +14,7 @@ func initJwtAndAllow(ctx *godog.ScenarioContext) {
 
 	scenario := istioJwtManifestScenario{s}
 
+	ctx.Step(`JwtAndUnrestricted: There is a httpbin service$`, scenario.thereIsAHttpbinService)
 	ctx.Step(`JwtAndUnrestricted: There is an endpoint secured with JWT on path "([^"]*)"$`, scenario.thereIsAnJwtSecuredPath)
 	ctx.Step(`JwtAndUnrestricted: There is an endpoint with handler "([^"]*)" on path "([^"]*)"$`, scenario.thereIsAnEndpointWithHandler)
 	ctx.Step(`JwtAndUnrestricted: The APIRule is applied$`, scenario.theAPIRuleIsApplied)
