+++ tests/integration/scenario_jwt_issuer_jwks_not_match_test.go
@@ -12,6 +12,7 @@ func initJwtIssuerJwksNotMatch(ctx *godog.ScenarioContext) {
 
 	scenario := istioJwtManifestScenario{s}
 
+	ctx.Step(`JwtIssuerJwksNotMatch: There is a httpbin service$`, scenario.thereIsAHttpbinService)
 	ctx.Step(`JwtIssuerJwksNotMatch: There is an endpoint secured with JWT on path "([^"]*)" with invalid issuer and jwks$`, scenario.emptyStep)
 	ctx.Step(`JwtIssuerJwksNotMatch: The APIRule is applied$`, scenario.theAPIRuleIsApplied)
 	ctx.Step(`JwtIssuerJwksNotMatch: Calling the "([^"]*)" endpoint with a valid "([^"]*)" token should result in body containing "([^"]*)"$`, scenario.callingTheEndpointWithValidTokenShouldResultInBodyContaining)
