+++ tests/integration/scenario_jwt_unavailable_issuer_test.go
@@ -12,6 +12,7 @@ func initJwtUnavailableIssuer(ctx *godog.ScenarioContext) {
 
 	scenario := istioJwtManifestScenario{s}
 
+	ctx.Step(`JwtIssuerUnavailable: There is a httpbin service$`, scenario.thereIsAHttpbinService)
 	ctx.Step(`JwtIssuerUnavailable: There is an endpoint secured with JWT on path "([^"]*)" with invalid issuer and jwks$`, scenario.emptyStep)
 	ctx.Step(`JwtIssuerUnavailable: The APIRule is applied$`, scenario.theAPIRuleIsApplied)
 	ctx.Step(`JwtIssuerUnavailable: Calling the "([^"]*)" endpoint with a valid "([^"]*)" token should result in body containing "([^"]*)"$`, scenario.callingTheEndpointWithValidTokenShouldResultInBodyContaining)
