+++ tests/integration/scenario_oauth_and_jwt_test.go
@@ -12,6 +12,7 @@ func initJwtAndOauth(ctx *godog.ScenarioContext) {
 
 	scenario := istioJwtManifestScenario{s}
 
+	ctx.Step(`OAuth2: There is a httpbin service$`, scenario.thereIsAHttpbinService)
 	ctx.Step(`OAuth2: There is an endpoint secured with JWT on path "([^"]*)" requiring scopes '(\[.*\])'$`, scenario.thereIsAnEndpointWithRequiredScopes)
 	ctx.Step(`OAuth2: There is an endpoint secured with OAuth2 on path "([^"]*)" requiring scopes '(\[.*\])'$`, scenario.thereIsAnEndpointWithRequiredScopes)
 	ctx.Step(`OAuth2: The APIRule is applied$`, scenario.theAPIRuleIsApplied)
