+++ tests/integration/scenario_service_fallback_test.go
@@ -12,6 +12,7 @@ func initJwtServiceFallback(ctx *godog.ScenarioContext) {
 
 	scenario := istioJwtManifestScenario{s}
 
+	ctx.Step(`ServiceFallback: There is a httpbin service$`, scenario.thereIsAHttpbinService)
 	ctx.Step(`ServiceFallback: There is an endpoint secured with JWT on path "([^"]*)" with service definition$`, scenario.thereIsAnEndpointWithServiceDefinition)
 	ctx.Step(`ServiceFallback: There is an endpoint secured with JWT on path "([^"]*)"$`, scenario.thereIsAnJwtSecuredPath)
 	ctx.Step(`ServiceFallback: The APIRule with service on root level is applied$`, scenario.theAPIRuleIsApplied)
